Adjusted existing classes to look for moved dependencies.
