+++ frontend/src/main/java/cz/cuni/mff/xrg/odcs/frontend/gui/dialog/DPUDetail.java
@@ -2,6 +2,10 @@ package cz.cuni.mff.xrg.odcs.frontend.gui.dialog;
 
 import java.io.FileNotFoundException;
 
+import cz.cuni.mff.xrg.odcs.commons.app.dpu.wrap.DPUInstanceWrap;
+import cz.cuni.mff.xrg.odcs.commons.app.dpu.wrap.DPUWrapException;
+import cz.cuni.mff.xrg.odcs.commons.app.facade.ModuleFacade;
+import cz.cuni.mff.xrg.odcs.frontend.AppEntry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.vaadin.dialogs.ConfirmDialog;
@@ -16,8 +20,6 @@ import cz.cuni.mff.xrg.odcs.commons.app.facade.DPUFacade;
 import cz.cuni.mff.xrg.odcs.commons.app.facade.RuntimePropertiesFacade;
 import cz.cuni.mff.xrg.odcs.commons.app.module.ModuleException;
 import cz.cuni.mff.xrg.odcs.frontend.auxiliaries.DecorationHelper;
-import cz.cuni.mff.xrg.odcs.frontend.dpu.wrap.DPUInstanceWrap;
-import cz.cuni.mff.xrg.odcs.frontend.dpu.wrap.DPUWrapException;
 import cz.cuni.mff.xrg.odcs.frontend.gui.components.DPUConfigHolder;
 import cz.cuni.mff.xrg.odcs.frontend.gui.components.DPUGeneralDetail;
 import cz.cuni.mff.xrg.odcs.frontend.i18n.Messages;
@@ -201,7 +203,8 @@ public class DPUDetail extends Window {
      * @param readOnly
      */
     public void showDpuDetail(DPUInstanceRecord dpu, boolean readOnly) {
-        this.dpuInstance = new DPUInstanceWrap(dpu, dpuFacade, runtimePropertiesFacade.getLocale());
+        ModuleFacade moduleFacade = ((AppEntry) UI.getCurrent()).getBean(ModuleFacade.class);
+        this.dpuInstance = new DPUInstanceWrap(dpu, dpuFacade, moduleFacade, runtimePropertiesFacade.getLocale());
         this.setCaption(Messages.getString("DPUDetail.detail", dpu.getName().trim(),
                 readOnly ? Messages.getString("DPUDetail.read-only.mode") : ""));
